make first question

+++ tests/test_q1.py
@@ -8,8 +8,8 @@ import pathlib
 from dataclasses import dataclass
 from dataclasses_json import dataclass_json
 
-BUFFER_SIZE=2500
-index = json.loads(pathlib.Path("./index.json").read_text())
+BUFFER_SIZE = 2500
+
 
 @dataclass_json
 @dataclass
@@ -27,32 +27,63 @@ class Product:
     manufacturer: Optional[str]
 
 
-@pytest.fixture(autouse=True)
+@pytest.fixture(scope="module", autouse=True)
 def populate_and_delete(elasticsearch_ready):
     # Create index
-    index = "my-index-000001"
+    index = "m-index-1"
     uri = f"{elasticsearch_ready}/{index}"
-    logging.info(uri)
-    resp = requests.put(uri, json=index)
+    settings = json.loads(pathlib.Path("./index.json").read_text())
+    while True:
+        resp = requests.put(uri, json=settings)
+        if resp.status_code == 200:
+            break
+        elif resp.status_code == 400:
+            requests.delete(uri)
+        else:
+            assert False, print(resp.text)
     # Populate index
-    num_lines = sum(1 for line in open('productos_cleaned.csv')) - 1
+    num_lines = sum(1 for line in open("productos_cleaned.csv")) - 1
     with open("productos_cleaned.csv", "r") as f:
-        reader = csv.DictReader(f, delimiter=',', quotechar='"')
+        reader = csv.DictReader(f, delimiter=",", quotechar='"')
         bulk_msg = []
         for i, row in enumerate(reader):
-            p = Product.from_dict({"id":i, **row})
-            bulk_msg.append(json.dumps({"index": {"_index":index, "_id": i}}))
+            p = Product.from_dict({"id": i, **row})
+            bulk_msg.append(json.dumps({"index": {"_index": index, "_id": i}}))
             bulk_msg.append(p.to_json())
-            if (i+1) % BUFFER_SIZE == 0 or i+1 == (num_lines) :
+            if (i + 1) % BUFFER_SIZE == 0 or i + 1 == num_lines:
                 bulk_msg.append("")
-                resp = requests.put(f"{uri}/_bulk", data='\n'.join(bulk_msg), headers={'content-type':'application/json'}).raise_for_status()
-                print("Data insterted {}/{} ({:.2f})".format(i+1, num_lines, (i+1)/num_lines))
+                resp = requests.put(
+                    f"{uri}/_bulk",
+                    data="\n".join(bulk_msg),
+                    headers={"content-type": "application/json"},
+                ).raise_for_status()
+                print(
+                    "Data insterted {}/{} ({:.2f})".format(
+                        i + 1, num_lines, (i + 1) / num_lines
+                    )
+                )
 
     yield uri
-    # Delete index and any document
-    requests.delete(uri)
 
-def test_insert_data(populate_and_delete, caplog):
+
+@pytest.mark.parametrize("query", sorted(pathlib.Path("./queries").glob("q1*.json")))
+def test_multi_language(populate_and_delete, query, caplog):
     with caplog.at_level(logging.DEBUG):
-        print(populate_and_delete)
-        assert True
+        q = json.loads(pathlib.Path(query).read_text())
+        resp = requests.get(f"{populate_and_delete}/_search", json=q)
+        resp.raise_for_status()
+        d = resp.json()
+        assert len(d["hits"]["hits"]) > 0, print(resp.text)
+        print()
+        print(
+            "Top results for query: '{}', terms: '{}'".format(
+                query, q["query"]["multi_match"]["query"]
+            )
+        )
+        print("\t".join(["id", "score", "description"]))
+        for hit in d["hits"]["hits"]:
+            print(
+                "\t".join(
+                    [hit["_id"], str(hit["_score"]), hit["_source"]["description"]]
+                )
+            )
